requirements.txt
pyjwt
passlib
pydantic
uvicorn
fastapi
sqlalchemy
psycopg2-binary
pydantic-settings
databases
python-dotenv
psutil
alembic
asyncpg
bcrypt
PyPDF2
pdfplumber
xlsxwriter
openpyxl
openai
google-generativeai